+++ apps/io-wallet-user-func/src/infra/azure/functions/create-wallet-unit-attestation.ts
@@ -1,0 +1,7 @@
+import { httpAzureFunction } from "@pagopa/handler-kit-azure-func";
+
+import { CreateWalletUnitAttestationHandler } from "@/infra/http/handlers/create-wallet-unit-attestation";
+
+export const CreateWalletUnitAttestationFunction = httpAzureFunction(
+  CreateWalletUnitAttestationHandler,
+);