+++ apps/io-wallet-user-func/src/infra/http/handlers/create-wallet-instance.ts
@@ -1,7 +1,6 @@
 import { QueueClient } from "@azure/storage-queue";
 import * as H from "@pagopa/handler-kit";
 import { FiscalCode, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
-import { sequenceS } from "fp-ts/Apply";
 import { flow, pipe } from "fp-ts/function";
 import * as E from "fp-ts/lib/Either";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
@@ -14,6 +13,7 @@ import {
   validateAttestation,
   ValidatedAttestation,
 } from "@/attestation-service";
+import { revokeAllCredentials } from "@/credential";
 import { enqueue } from "@/infra/azure/storage/queue";
 import { sendTelemetryExceptionWithBody } from "@/telemetry";
 import { isLoadTestUser } from "@/user";
@@ -41,21 +41,20 @@ const requireWalletInstanceRequest = (req: H.HttpRequest) =>
   pipe(
     req.body,
     H.parse(WalletInstanceRequestPayload),
-    E.chain(
+    E.map(
       ({
         challenge,
         fiscal_code,
         hardware_key_tag,
         is_renewal,
         key_attestation,
-      }) =>
-        sequenceS(E.Apply)({
-          challenge: E.right(challenge),
-          fiscalCode: E.right(fiscal_code),
-          hardwareKeyTag: E.right(hardware_key_tag),
-          isRenewal: E.right(is_renewal ?? false),
-          keyAttestation: E.right(key_attestation),
-        }),
+      }) => ({
+        challenge: challenge,
+        fiscalCode: fiscal_code,
+        hardwareKeyTag: hardware_key_tag,
+        isRenewal: is_renewal ?? false,
+        keyAttestation: key_attestation,
+      }),
     ),
   );
 
@@ -98,6 +97,29 @@ export const CreateWalletInstanceHandler = H.of((req: H.HttpRequest) =>
                   ? "WALLET_INSTANCE_RENEWAL"
                   : "NEW_WALLET_INSTANCE_CREATED",
               ),
+            ),
+            RTE.chainW((hasRevokedOldWalletInstances) =>
+              hasRevokedOldWalletInstances
+                ? pipe(
+                    walletInstanceRequest.fiscalCode,
+                    revokeAllCredentials,
+                    // If revokeAllCredentials fails, emit telemetry and continue revoking the wallet instance
+                    // This avoids blocking wallet revocation on external dependencies
+                    RTE.orElseW(
+                      flow(
+                        sendTelemetryExceptionWithBody({
+                          body: req.body,
+                          functionName: "createWalletInstance",
+                        }),
+                        E.fold(
+                          () => E.right(undefined),
+                          () => E.right(undefined),
+                        ),
+                        RTE.fromEither,
+                      ),
+                    ),
+                  )
+                : RTE.right(undefined),
             ),
             RTE.chainW(() =>
               walletInstanceRequest.isRenewal