+++ apps/io-wallet-user-func/src/infra/mobile-attestation-service/android/assertion.ts
@@ -59,7 +59,7 @@ export const verifyAssertion = async (
   const errors: string[] = [];
 
   // First check whether the clientData has been signed correctly with the hardware key
-  const signatureValidated = validateAssertionSignature(
+  const signatureValidated = await validateAssertionSignature(
     hardwareKey,
     clientData,
     hardwareSignature,
@@ -156,11 +156,9 @@ export const validateAssertionSignature = async (
     throw new Error("Invalid Hardware Key format");
   }
   const publicHardwareKeyPem = await exportSPKI(joseHardwareKey);
-
-  const clientDataHash = createHash("sha256").update(clientData).digest();
-
   const verifier = createVerify("SHA256");
-  verifier.update(clientDataHash);
+  verifier.update(clientData);
+
   return verifier.verify(publicHardwareKeyPem, hardwareSignature, "base64");
 };