+++ apps/io-wallet-user-func/src/verifier.ts
@@ -1,5 +1,4 @@
 import { parse } from "@pagopa/handler-kit";
-import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import * as E from "fp-ts/Either";
 import { pipe } from "fp-ts/function";
 import * as TE from "fp-ts/TaskEither";
@@ -18,7 +17,7 @@ type WithJwkCnf = t.TypeOf<typeof WithJwkCnf>;
 const decodeJwt = (jwt: string) =>
   E.tryCatch(() => jose.decodeJwt(jwt), E.toError);
 
-export const getPublicKeyFromCnf = (jwt: string) =>
+const getPublicKeyFromCnf = (jwt: string) =>
   pipe(
     jwt,
     decodeJwt,
@@ -31,7 +30,7 @@ export const getPublicKeyFromCnf = (jwt: string) =>
     E.map((payload) => payload.cnf.jwk),
   );
 
-export const verifyAndDecodeJwt = (jwt: string) => (publicKey: JwkPublicKey) =>
+const verifyAndDecodeJwt = (jwt: string) => (publicKey: JwkPublicKey) =>
   pipe(
     TE.tryCatch(() => jose.importJWK(publicKey), E.toError),
     TE.chain((joseKey) =>
@@ -52,45 +51,3 @@ export const verifyJwtWithInternalKey = (jwt: string) =>
     TE.fromEither,
     TE.chain(verifyAndDecodeJwt(jwt)),
   );
-
-interface WithHeaderKidAndPayloadCnfJwkKid {
-  header: {
-    kid: string;
-  };
-  payload: {
-    cnf: {
-      jwk: {
-        kid: string;
-      };
-    };
-  };
-}
-
-interface WithIssuerAndHardwareKeyTag {
-  hardwareKeyTag: NonEmptyString;
-  iss: NonEmptyString;
-}
-
-export const validateIssuerMatchesHardwareKeyTag = <
-  T extends WithIssuerAndHardwareKeyTag,
->(
-  jwt: T,
-): E.Either<Error, void> =>
-  jwt.iss === jwt.hardwareKeyTag
-    ? E.right(undefined)
-    : E.left(
-        new Error(
-          "Invalid jwt: payload.iss must match payload.hardware_key_tag",
-        ),
-      );
-
-export const validateHeaderKidMatchesCnfKid = <
-  T extends WithHeaderKidAndPayloadCnfJwkKid,
->(
-  jwt: T,
-): E.Either<Error, void> =>
-  jwt.payload.cnf.jwk.kid === jwt.header.kid
-    ? E.right(undefined)
-    : E.left(
-        new Error("Invalid jwt: header.kid must match payload.cnf.jwk.kid"),
-      );