+++ infra/resources/psn/prod/key_vault_secrets.tf
@@ -32,3 +32,38 @@ resource "azurerm_key_vault_secret" "application_insights_connection_string" {
   value_wo         = data.azurerm_application_insights.core.connection_string
   value_wo_version = 1
 }
+
+resource "azurerm_key_vault_secret" "user_ioweb_fn_key" {
+  name             = "user-ioweb-fn-key"
+  key_vault_id     = module.key_vault_app.key_vault_wallet.id
+  value_wo         = ""
+  value_wo_version = 1
+}
+
+resource "azurerm_key_vault_secret" "user_ioapp_fn_key" {
+  name             = "user-ioapp-fn-key"
+  key_vault_id     = module.key_vault_app.key_vault_wallet.id
+  value_wo         = ""
+  value_wo_version = 1
+}
+
+resource "azurerm_key_vault_secret" "user_uat_fn_default_key" {
+  name             = "user-fn-uat-default-key"
+  key_vault_id     = module.key_vault_app.key_vault_wallet.id
+  value_wo         = ""
+  value_wo_version = 1
+}
+
+resource "azurerm_key_vault_secret" "support_fn_default_key" {
+  name             = "support-fn-default-key"
+  key_vault_id     = module.key_vault_app.key_vault_wallet.id
+  value_wo         = ""
+  value_wo_version = 1
+}
+
+resource "azurerm_key_vault_secret" "application_insights_connection_string" {
+  name             = "ApplicationInsightsConnectionString"
+  key_vault_id     = module.key_vault_app.key_vault_wallet.id
+  value_wo         = data.azurerm_application_insights.core.connection_string
+  value_wo_version = 1
+}